+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 Nothing yet.
 
+### Fixes
+
+* RequireAnyOneOf - now accepts callables for requirements (instead of requiring only callable objects)
+
 ## 1.0.0 - Wed Jul 22 2015
 
 Initial release. Largely based on Stuart's old ContractLib2.